final version

+++ srcs/requirements/wordpress/tools/script_wp.sh
@@ -1,51 +1,20 @@
 #!/bin/bash
 
-# create directory to use in nginx container later and also to setup the wordpress conf
-mkdir /var/www/
-mkdir /var/www/html
+sleep 10
 
-cd /var/www/html
+if [ ! -f /var/www/html/wp-config.php ]; then
+    echo "WordPress is installing ..."
+    cd /var/www/html
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+    wp --allow-root --path=/var/www/html/ core download
+    wp --allow-root --path=/var/www/html/ config create --dbname=$db_name --dbuser=$db_user --dbpass=$db_pwd --dbhost=mariadb:3306
+    wp --allow-root --path=/var/www/html/ core install --url=$DOMAIN_NAME --title=$WP_TITLE --admin_name=$WP_MOD_USR --admin_password=$WP_MOD_PWD --admin_email=$WP_MOD_EMAIL
+    wp --allow-root user create $WP_USR $WP_EMAIL --role=author --user_pass=$WP_PWD
+    chown -R www-data:www-data /var/www/html
+else
+    echo "WordPress is already installed!"
+fi
 
-rm -rf *
-
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
-
-chmod +x wp-cli.phar 
-
-mv wp-cli.phar /usr/local/bin/wp
-
-wp core download --allow-root
-
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-
-# change the those lines in wp-config.php file to connect with database
-
-sed -i -r "s/database/$db_name/1"   wp-config.php
-sed -i -r "s/database_user/$db_user/1"  wp-config.php
-sed -i -r "s/passwod/$db_pwd/1"    wp-config.php
-
-sed -i -r "s/localhost/mariadb/1"    wp-config.php  (to connect with mariadb database)
-
-# installs WordPress and sets up the basic configuration for the site. The --url option specifies the URL of the site, --title sets the site's title, --admin_user and --admin_password set the username and password for the site's administrator account, and --admin_email sets the email address for the administrator. The --skip-email flag prevents WP-CLI from sending an email to the administrator with the login details.
-wp core install --url=$DOMAIN_NAME/ --title=$WP_TITLE --admin_user=$WP_ADMIN_USR --admin_password=$WP_ADMIN_PWD --admin_email=$WP_ADMIN_EMAIL --skip-email --allow-root
-
-# creates a new user account with the specified username, email address, and password. The --role option sets the user's role to author, which gives the user the ability to publish and manage their own posts.
-wp user create $WP_USR $WP_EMAIL --role=author --user_pass=$WP_PWD --allow-root
-
-# active theme for the site astra
-wp theme install astra --activate --allow-root
-
-
-wp plugin install redis-cache --activate --allow-root
-
-
-# uses the sed command to modify the www.conf file in the /etc/php/7.3/fpm/pool.d directory. The s/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/g command substitutes the value 9000 for /run/php/php7.3-fpm.sock throughout the file. This changes the socket that PHP-FPM listens on from a Unix domain socket to a TCP port.
-sed -i 's/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/g' /etc/php/7.3/fpm/pool.d/www.conf
-
-# creates the /run/php directory, which is used by PHP-FPM to store Unix domain sockets.
-mkdir /run/php
-
-
-wp redis enable --allow-root
-
-/usr/sbin/php-fpm7.3 -F
+exec php-fpm8.2 -F
